Added explicit lein dependency in install section of readme
The installation instructions in the readme.md file don’t explicitly say to add midje to the project dependencies. Just starting out with lein-midje, I stumbled around quite a bit trying to get it installed. I added a mention of this dependency to the installation section along with an alternative :profile that works in the case where someone has not yet added a dependency on midje. (Credit where due, I found the solution here: http://stackoverflow.com/tags/leiningen/new
Anyhow, there may be a better way to do it, but these instructions would have made my first encounter with lein-midje much smoother.
Also, while I was at it, I brought the lein-midje version up to 3.1.3, which seems to be current.